HttpGet Actions By UserId

+++ PresentsYouAPI/Models/Education.cs
@@ -6,7 +6,7 @@
         public string Institute { get; set; }
         public string Degree { get; set; }
         public string Field { get; set; }
-        public int Grade { get; set; }
+        public float Grade { get; set; }
         public int UserId { get; set; }
     }
 }
